Implements functional BLE relay
